+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 ## New Features
 
+- Add `var::ContainerObject` as the base to container classes
+  - Provides more access to algorithms. Allows for easier iteration.
 - Add `api::SystemResource` to encapsulate system resources and minimize the use of the rule of 5.
 - Documentation updates for `API`
 - Use a `printer::Printer*` to print help when using `sys::Cli::show_help()`